+++ src/module/data/item/career.mjs
@@ -2,6 +2,10 @@ import { systemPath } from "../../constants.mjs";
 import AdvancementModel from "./advancement.mjs";
 
 /**
+ * @import CharacterModel from "../actor/character.mjs";
+ */
+
+/**
  * Careers describe what a hero did for a living before becoming a hero
  */
 export default class CareerModel extends AdvancementModel {
@@ -31,5 +35,41 @@ export default class CareerModel extends AdvancementModel {
     schema.wealth = new fields.NumberField({ integer: true, required: true });
 
     return schema;
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  _onCreate(data, options, userId) {
+    super._onCreate(data, options, userId);
+    if ((userId !== game.userId) || !this.actor) return;
+
+    /** @type {CharacterModel} */
+    const systemModel = this.actor.system;
+
+    this.actor.update({
+      system: {
+        wealth: systemModel.hero.wealth + this.wealth,
+        renown: systemModel.hero.renown + this.renown,
+      },
+    });
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  _onDelete(options, userId) {
+    super._onDelete(options, userId);
+    if ((userId !== game.userId) || !this.actor) return;
+
+    /** @type {CharacterModel} */
+    const systemModel = this.actor.system;
+
+    this.actor.update({
+      system: {
+        wealth: systemModel.hero.wealth - this.wealth,
+        renown: systemModel.hero.renown - this.renown,
+      },
+    });
   }
 }