+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Advancements
 - Added a button to ability use messages to apply effects from the ability. (#214)
 - Added a Wealth field to Careers. (#535)
+- Adding or removing a career will adjust the amount of renown and wealth your character has.
 - Added a Renown and Wealth field to characters. (#279)
 - Added age, height, and weight info to the Biography tab for characters.